Adds dashboard image carousel to barangay info
Enables administrators to manage a rotating carousel of images on the dashboard.

This enhances the dashboard with visually appealing content, such as images of barangay landmarks, events, or announcements.

Includes changes to the BarangayInfo entity, DTO, use case, repository, and database schema to support the management and storage of dashboard image paths.

Also renames `example/profiles` images to `example/officials`.